Create color.h
I was messing with this today instead of working on lz4 heh. the color struct can handle conversion between different color formats on the fly. should work for any bits per channel. tested with 16, 8 and 5.

my issue with lz4 is I need to detect the compressed size to pass only the data i need to decompress. I think this might be in the first 12 bytes of the header. but i need to double check. might debug on openviii-monogame and look at those bytes to see what's there.