support for std::string in place of std::vector is working using templates.
There are alot of template to make this work.

Some templates could be replaced with span in cpp20. The templates allow us to handle a vector or a sting. Techinically you should be able to move a vector into a string as both values are contiguous chars. But this is no support for this in the std::library. I am working to being able to std::forward or std::move the data into the FIFLFS so we are not coping the data. Basically if you don't move you are coping.