assigning data directly to the FIFLFS. instead of passing a copy of the data.
This required some reordering of code. FS's static function can now read from a zzz file because I needed to do that for it to work.  FileData has some extra functions to return something the FS code expects. This isn't the fastest way to dump data from the archives but that isn't the point it's to test the functions. and make sure things are working. if I wanted to dump I'd almost never store the data in ram I'd put it directly into an output file. I might return a file handler. in a future build. So I can try to read and write in a loop instead of using ram. The movie and music player would work like this as we wouldn't be reading the entire contents into ram we'd just be filling a buffer and then waiting till it emptys.