trying to setup some things via compile time methods
now i'm gonnna setup TestArchives to see if my code is working.

After assuring the FIFLFS files are read populating. I want to make sure ZZZ files populate and also grab the FIFLFS from the ZZZ file.

This will require setting a language code. The steam version has a file with the language code in it. the remaster stores this in the documents folder I think. near impossible to get that because of the unique identifer.