Mostly Kernel updates.
Kernel is kinda huge. Alot of data to parse. Each with a slightly different layout ;-;

I am trying to make sure the objects with strings parse correctly.  There are some strings in namedic.bin that the menus and kernel uses. I need to support that so kernel can reference those strings. Though it's been a while since I looked at the strings in that file. They have to be embedded before the kernel strings are converted to utf8. like a prepass. So might have a class/struct like FF8String that does a similar actions but instead of having strings hard coded they will be loaded and parsed using spans/string-view of char.