loaded up msvc and tried to get rid of some warnings that popped up.
It didn't like my strings being returned said they should be const. So i made constexpr varibles and returned those...

I also found how to reformat all my files at once so ran it on the whole tree.

and msvc compained i was using unicode characters in a non unicode string so added u8 to start.

added elses because it compained about unreachable code. Though else after a return shouldn't be needed as it would go there if and only if the if statement is false. Though I guess if I'm using if constexpr then I'm generating functions with code that will never be reached. By adding the else the code will be excluded from the function instead of being skipped code.