added a std::variant alternative
you can't really choose an archive at runtime with a no constexpr value. So these new functions will let you grab a std::varient of the possible archive types. It's probably not something I'd want to use in the day to day. I'm mostly trying to think how to use my search results...
atleast the search results are fast. Just running code on everything seemed slow.

I'll probably see about maybe making a copy of my search function and using some kinda function pointer. or getting a tuple of the file buffer out of the results so i can do something with it. I cannot really do anything with the number because I need to reload the archive and pass the number to it to get the fi and then request the buffer using the fi. AHHH. :P

Though I could use the filename maybe Though that would be silly And I'd still need to know which file archive the file is in. So I'd rather get a buffer and a filename as a bare minumum so i can run code.

Though I also don't want to store all those buffers in ram. So I think it's better to pass a lamda or a function pointer to the archive and have it execute the code on the buffer and then return the results in a... I donno.