reswizzle test is working
Now I need to take thc code to reswizzle and make it into a class or something. And then write a test that deswizzles the reswizzle to test to see if it matches the deswizzle from the raw images.

ecenter3 had two maps and mim files i had to be explict about the filenames instead of only checking the extensions