progress on deswizzle reswizzle
changed deswizzle output buffer to 24 bit color.  It will accept a generic color. if it's not the same type it will force a conversion.
minium interface functions for MimFromPath. Now I need to read the textures into a buffer I can query.
fixed case in Color24 conversion. As it was uppercase and the functions were renamed to lowercase.
Moved some common code for get_raw_width into a static function.