Update FI.hpp
I almost think the FI shouldn't have a constructor that uses span or istream. Because it is a trivial type. I can just read it from the buffer. Though one thing I would need to check is the offset

for now I've updated the constructor to use Tools::read_val<> which does all the memcpy for you.