fixed the output for nested values.
okay somewhere it's leaving the \r on the end of the file path heh. I thought I was filtering that out but must of missed a spot.
[8:20 AM]
so when i was saving the files the string printed out was empty  because \r goes back to the front overwrites everything on the line.

SebanisuToday at 8:40 AM
hmm so it was letting my compile passing a const variable to the clean buffer method. So the cleanbuffer function didn't do anything. Maybe it was making a copy of the string when i passed the const to it.