Update Read.hpp
I needed to be able to pass 0 size and it not memcpy anything. Previous behavior it would fill the container with max number of objects if size was 0.

So i had to make size not optional and have a second functiont that doesn't take a size that would compute the max number of objects.