fiflfs extract complete (#45)
It might need updates but it's working. The new app extracts an archive to a folder.

the .compress file has a byte for each file saying if it's compressed or not. and the .names has the paths in the same order. it's one of the easiest ways I could think of doing it.