Update FIFLFS.hpp
added a function that lets us take files that are in a vector and search for matches. in the archive and replace the files writing to the temp directory. Gives us a new vector with the paths to the new files. so after we processed all of fields we can combine the vectors and create a new archive with the new archives.